Refactor: split `index.js` into lib files
Moves code around without changing behavior, but it does fix a
memory leak in `db.clear()` (it previously did not close the
iterator).

Category: change
