+++ package.json
@@ -15,6 +15,7 @@
   "files": [
     "index.js",
     "index.d.ts",
+    "lib",
     "UPGRADING.md",
     "CHANGELOG.md"
   ],
